// src/app/pages/characters/models/character.interface.ts
export interface CharactersApiResponse {
  info: Info;
  results: CharacterEntity[];
}

export interface Info {
  count: number;
  pages: number;
  next: string;
  prev: any;
}

export interface CharacterEntity {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: Origin;
  location: Location;
  image: string;
  episode: string[];
  url: string;
  created: string;
}

export interface Origin {
  name: string;
  url: string;
}

export interface Location {
  name: string;
  url: string;
}